API Integration
The create a new account page now takes you to login again. Added a SingletonQueue which can handle API calls. The CocktailDataService uses the singleton to call and pass info too the Search activity which now displays the name and image of a random cocktail. Would be good to add a way to view ingredients, but we should prioritize authenticating a login first.